+++ apps/webapp/app/services/gitHubAuth.server.ts
@@ -5,6 +5,7 @@ import { createFirstOrganization } from "~/models/organization.server";
 import { findOrCreateUser } from "~/models/user.server";
 import type { AuthUser } from "./authUser";
 import { scheduleWelcomeEmail } from "./email.server";
+import { taskQueue } from "./messageBroker.server";
 
 const gitHubStrategy = new GitHubStrategy(
   {
@@ -31,6 +32,16 @@ const gitHubStrategy = new GitHubStrategy(
       if (isNewUser) {
         await createFirstOrganization(user);
         await scheduleWelcomeEmail(user);
+
+        await taskQueue.publish("SEND_INTERNAL_EVENT", {
+          id: user.id,
+          name: "user.created",
+          payload: {
+            id: user.id,
+            source: "GITHUB",
+            admin: user.admin,
+          },
+        });
       }
 
       return {