+++ apps/webapp/app/utils/formData.ts
@@ -1,0 +1,10 @@
+import type { z } from "zod";
+
+export function formDataAsObject<TValues>(
+  formData: FormData,
+  schema: z.ZodSchema<TValues>
+): TValues {
+  const object = Object.fromEntries(formData.entries());
+  const parsed = schema.parse(object);
+  return parsed;
+}