+++ apps/webapp/app/utils/objects.ts
@@ -1,0 +1,14 @@
+export function omit<T extends Record<string, unknown>, K extends keyof T>(
+  obj: T,
+  keys: K[]
+): Omit<T, K> {
+  const result: any = {};
+
+  for (const key of Object.keys(obj)) {
+    if (!keys.includes(key as K)) {
+      result[key] = obj[key];
+    }
+  }
+
+  return result;
+}