+++ apps/webapp/prisma/migrations/20230127121743_add_slack_interaction_trigger_type/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "TriggerType" ADD VALUE 'SLACK_INTERACTION';